fix(extensions): shared/ の ESM scope 復元で Playwright e2e の CJS 誤判定を解消 (#1586)
CI の check job（suno-helper e2e）が SyntaxError: Named export not found
（CommonJS module）で全滅していた。根本原因は #1629（ADR-0021）の root
package.json（"type": "module"）削除で、extensions/shared/ 配下の .ts が
Node ESM loader（Playwright の TS transform）に CJS スコープ判定される
ようになったこと（main では path filter により extensions CI が走らず潜伏）。

- extensions/shared/package.json（type: module, private）を新設しスコープを復元
- #1586 が e2e に追加していた queue runner / completion gate の純ロジック
  テスト 3 件（page 不使用）を tests/queue-runner.test.ts（vitest）へ移設し、
  suno-queue.spec.ts は実ブラウザ layout を要する DOM スモークに限定
  （queue-runner モジュールの直接ユニットテスト不足の指摘にも対応）

検証: suno-helper vitest 970 / e2e 27、distrokid-helper vitest 205 / e2e 6、
tsc --noEmit、prettier すべて通過。
