+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `fix(suno-helper)`: queue 実行の resume payload 境界で `submittedClipIds` / `playlistExpectedClipCount` を fail-loud に検証し、未完了 clip を保持した再開時に不正な raw payload が playlist 完了待ちへ流れ込まないようにした（#1586）。
+- `fix(extensions)`: suno-helper の Playwright e2e が `SyntaxError: Named export ... not found（CommonJS module）` で全滅していた CI 失敗を修正（#1586 ブランチで顕在化）。根本原因は #1629（ADR-0021）の root `package.json`（`"type": "module"`）削除で `extensions/shared/` が Node ESM loader に CJS スコープ判定されるようになったこと。`extensions/shared/package.json`（`type: module`）を新設してスコープを復元した。あわせて #1586 が e2e に追加していた queue runner / completion gate の純ロジックテスト 3 件（page 不使用）を `tests/queue-runner.test.ts`（vitest）へ移設し、e2e は実ブラウザ layout を要する DOM スモークに限定した
 - `fix(suno-helper)`: reload 後の resume / playlist 再実行で、保存済み clip が照会時点で complete 済みだと完了待ちゲートが永遠に pending 扱いして 10 分の stall timeout まで空回りし偽 ERROR で停止する回帰を修正（#1586 レビュー指摘）。active feed poll（`FEED_V3_POLL_RESPONSE`）の応答は `ClipTracker.applyRequestedStatuses` で未知+終端 clip も登録する（passive `FEED_CLIPS` の「未知+終端は捨てる」規則は据え置き）。あわせて同レビューの指摘を反映: `retryPlaylist` の完了待ち中に popup が `[1/0] 注入中` を表示し続ける問題へ待機中 progress を追加、queue runner の中断分岐が `isEntrySubmitted` へ `EntryRunResult` を誤って渡していた型不整合をシグネチャ縮小（`(index) => boolean`）で解消、no-op だった `resolveQueueFatalInterruptIndex` を削除、entry 表示名 fallback（`title ?? name`）を `entryDisplayName` に一本化、payload 検証で `-0` を非負整数/数として拒否、run mode storage I/O の未処理 rejection に warn fallback を追加、実行中は投入方式ラジオを無効化、`RUN_MODE_ORDER` を `RUN_MODES` のキーから導出、queue 投入ループの冗長な固定 25ms sleep を除去、`content-run-preflight` の `waitForQueueSlot` 既定モックを実装呼び出しへ戻し空 queue 回帰テストの空洞化を解消
 - `fix(skills)`: lifecycle skill の起動文字列を `bunx tayk <cmd>` から `uv run yt-<cmd>` へ戻した（#1625）。tayk は次世代版 CLI で現時点では運用未使用（実装は別リポジトリへ分離予定）のため、#965 で置換した表記のままではスキルを読んだ AI 実行者が未実装 subcommand（suno: `video-analyze` / `suno-verify`、masterup: `suno-audio-cleanup` / `suno-verify-playlist` / `fix-timestamps` / `finalize-master` / `apply-rain-layers` / `suno-select-tracks` ほか）で実行不能になっていた。対象は masterup / suno / suno-lyric / suno-helper / thumbnail / video-upload / video-description / wf-new / wf-next / playlist / analytics-collect / distrokid-helper の SKILL.md・references・config.default.yaml（wf-new が references として同梱する `yt-init-collection` の復旧ヒント文言も同様に戻した）。起動文字列のみの置換で、周辺の手順・節構成・TS 実装（packages/）は変更しない。回帰テストは方針を反転し、tayk cutover まで lifecycle skill に `tayk` 表記が入らないことを担保する `tests/test_lifecycle_skills_no_tayk.py` へ改称・書き換えた
 - `fix(skills)`: 音楽制作系 7 スキル（suno / suno-lyric / suno-helper / masterup / lyria / videoup / loop-video）の整合性監査で検出した記述ドリフトを修正（#1432）。`suno/SKILL.md` の slug 自動実行 argv 例に残っていた `["uv", "run", "bunx tayk video-analyze", ...]` の混在表記を `["bunx", "tayk", "video-analyze", ...]` に統一、`suno/references/suno-examples.md` から実装に存在しない設定キー `banned_adjective_free_instruments` への言及を削除、`loop-video/SKILL.md` の `generate_videos.sh` バージョン表記（v11.0 → v14）とループ背景の正規化 CRF 表記（20 → 22）を現行実装に合わせた
